movies_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2  # Конфигурируемая панель с информацией для отладки проекта
factory_boy==3.2.0
ipython==7.24.1
django_extensions==3.1.3
